Lock hmap access in dpif_netdev_flow_dump_next
Straightforward approach is to add locking around the hmap accessor
function only, ignoring potential modifications to the hmap during a
flow dump.  This appears to be equivalent to what OvS does with the
Linux kernel interface.

I am not sure what's intended to protect the actual flow itself
from being deleted while we have a reference to it.