Locking improvements for fast / slow path
- Introduce new dp_netdev_lookup_flow_locked and move the locking into
dp_netdev_lookup_flow.  Functions that already have the table_mutex
locked call the _locked version.

- Avoid the potential of the flow being deleted while we have a
reference to it by holding the lock in dp_netdev_port_input over flow
lookup and flow use.

- Write to the pipe after slow path output, and add the pipe's fd to
the poll() in the fast path.  This way the fast path will exit from and
re-enter poll() so that netmap transmits the packets put on the ring by
the slow path.

- Introduce a sanity check on the fds array in dp_thread_body.  The
existing XXX comment on the size of the array still applies.

- Avoid write() with an uninitialized value (even though it's a dummy
write).

XXX Right now it's not clear to me how the netmap writes from the slow
and fast path are protected from each other.